Move template and template generator.
